refactor : 유저와 딜러가 상속받는 추상화 클래스 플레이어 생성
