+++ src/main/kotlin/blackjack/domain/OutputInterface.kt
@@ -1,14 +1,14 @@
 package blackjack.domain
 
-import blackjack.domain.user.User
+import blackjack.domain.user.Player
 
 /**
  * view와 domain의 결합을 없애기 위한 interface
  * Created by Jaesungchi on 2022.06.15..
  */
 interface OutputInterface {
-    fun drawUserCard(user: User)
-    fun drawMoreCard(user: User)
+    fun drawUserCard(user: Player)
+    fun drawMoreCard(user: Player)
 
     fun drawDealerHitMessage()
 }
